backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-anthropic==0.2.4

# MCP (Model Context Protocol)
mcp==1.0.0

# Performance Testing APIs
lighthouse==0.1.0
playwright==1.40.0
selenium==4.16.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# PDF Generation
reportlab==4.0.7
pypdf2==3.0.1
weasyprint==60.2

# Screenshot & Image Processing
pillow==10.1.0
playwright==1.40.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Database (optional - for production)
redis==5.0.1
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx-test==0.21.0
